Finally, 2D_MT works perfectly nicely with ASCII I/O. Both sensTest2D (linear and log) and invTest2D (log) run and produce reasonable results. "Randie Mackie's format" now supports LOGE: add this string at the end of the first line to enforce natural log of resistivity. Matlab scripts will also understand this.
AIR LAYERS ARE STILL HARD CODED IN FORTRAN AND MATLAB I/O ROUTINES.

This is a major problem, since currently the origin is at the top of
the air layers. Which means that the vertical components of the
site locations have to be FIXED at 295240 meters (given the way
the air layers are currently computed in mkCond2D.m, in readCond_2D.m
and in the Fortran ioAscii.f90 module). Otherwise, the sites are
deep in the Earth or up in the air, which results in weird responses
and is extremely hard to debug - until you figure this out.

This will need to be updated, but everything currently works as it is,
as long as the site locations are correct. This is good enough for
our testing purposes at the moment.