Fixed the error that caused the segmentation fault for NLCG. The problem was that the cvector Dilu was deallocated through a deallocate() statement in Deallocate_Dilu(), rather than using a deall_cvector (which unsets Dilu%allocated too).
Here's an extract from my e-mail to Gary:

"I only get a segmentation fault with the Intel compiler (the executable compiled with g95 
just runs... but obviously the problem is still there). I started working on the 3D 
inversion in August (revision 72), and found out that the inversion exits after the first 
computation of the functional with a message

 Ccond or Econd in EdgeCond not allocated yet

I now remember that it took me some time to figure out what was going on. But adding this 
line

   modelDataInitialized = .false.

to exitSolver seemed to solve the problem. Indeed, we called ModelDataCleanUp and 
ModelOperatorCleanUp, but we didn't unset modelDataInitialized (and that had to cause 
problems). So I added the line (among lots of other things) and forgot about it, even 
though the inversion did act weird the first few iterations. I never had time to debug 
that.

Now, this is exactly the line that causes the segmentation fault with the Intel compiler. 
If I comment it out, I get the old EdgeCond message. This is something to do with the fact 
that we don't want to clear the model operators before computing the gradient, but we do 
want to clear them before the next iteration of NLCG. The logical thing to do would be to 
move these lines into the if(initForSens) loop (i.e. only run the deallocation subroutines 
if initFotSens == .true.). I tried that; the first gradient computation runs just fine in 
this case, but that causes a segmentation fault just before the following forward 
computation.

There either has to be an obvious solution to this, or this whole initForSens and 
modelDataInitialized business doesn't work in the context of an inversion! This is all a 
little different in 2D, so it's not easy to compare. I'll certainly try to debug this 
myself, but somehow I think it would be much more efficient if you could help, assuming you 
get this tomorrow morning (since we still need time to run the inversion!). Other than 
this, I have everything pretty much ready for both the synthetic inversion and the Dublin 
test.

The version that's in the SVN right now works just fine except for this problem that I 
described, so you could update it and try running it. All the input files should be in 
examples/ascii, and you could test it with

./Mod3DMT -I pr.cpr de.imp mi.cpr di.imp

Also, I should probably warn you that the 2D won't compile right now because of the few 
lines I inserted in NLCG to output a model every iteration (I have not restructured the 2D 
to have the I/O inside the ModelParameter yet, and it lacks the use statement). You could 
comment these lines out to compile the 2D if you wish."