Couple of important bug fixes and temporary patches to do with the user interface in 3D MT only. Details below.
Bug fixed in Grid3D that enables correct usage of the origin in 3D MT. For some reason, in Randie Mackie's model format the origin is supposedly in km, while the other grid definitions are in metres (at least that's what Gary's comment implies). For now, we stick with this arrangement: the origin has to be specified in km.

I have also corrected the units of the transfer functions, which were previously *wrong*. Here is the story. There are three kinds of units that are commonly used for the impedances:

(1) Ohm == [V/m]/[A/m]   SI units for E/H
(2) [V/m]/[T]            SI units for E/B
(3) [mV/km][nT]          practical units for E/B

Now, the modular code uses (2) internally (that is, SI units, and B rather than H). The conversion is like this:

1 Ohm = 104/(4*pi) [mV/km][nT] ~ 796 [mV/km][nT]
1 [mV/km][nT] = 103 [V/m]/[T]

So if the input is in Ohms, we need to multiply it by 107/(4*pi) in the modular code. If it is in [mV/km][nT], we should multiply it by 103.

If we are going to support all these options for the input, we should also support them for the output (so that the output has the same units as the input), - that would make it easier to compare the two. So the output routine should divide by these factors to get the units that the user has specified. This is how I implemented it for now.

Also, I have patched everything together to include the meta data in the data files: the headers that describe the data components and the site IDs. This should work for now. The generality of this program should make simple things like that even simpler; currently the poor interaction between the dictionaries and the data vectors in DataSpace makes even a simple modification like that exceedingly painful.  For example, if you wanted to gather all this information together to e.g. print out an intermediate solution (transfer functions) during an inversion, how would you go about that? 

I have a strong opinion about all this. I believe the whole dictionary / data structures and their interactions have to be completely re-thought. And re-done. The DataSpace module could only contain the data array, with no references to the transmitter or receiver. But in this case there should be an intermediate data structure that contains the data vector and the full transmitter and receiver information. This data structure should be used as the main source of transmitter / receiver information; and should also be used to set up the dictionaries, if they are ever needed. It needs thought and discussion, of course. But what is currently implemented in illogical, impractical and almost impossible to untangle! This is what makes this program so unnecessarily complicated and so difficult to debug. Structural independence of modules could (and should!) be implemented without such an overhead.