Added a structure to store the full sensitivity matrix to SensMatrix module. This is a cross between a data vector and a model parameter.
Ideal to store the full matrix J = df/dm
 OR
the gradient of the data misfit, component-wise, i.e. (J^T)_i r_i.
In this case, one column of J^T is a model parameter, so that J^T acts
on a single data residual to produce the misfit gradient for that data
point with respect to the model parameter; these can be summed up to
obtain the full gradient. Can be useful as another measure of sensitivity.

These matrices will be huge. We don't create them unless they are needed.

(Currently the sensitivity matrix in 2D & 3D MT is stored as a vector of
model parameters; taking up just as much space but not easy to make use of!)