Added some careful safeguarding in the global model parameter module.
Here's the comment for the linear combination routine:

  ! * New rules of the game (HIGHLY RECOMMENDED):
  ! * demand that the output isn't allocated when we call this and similar
  ! * routines. This prevents overwriting an input with a call such as
  ! * m1 = a1*m1 + a2*m2.
  ! * Calls like this cause errors that are extremely hard to debug.
  ! * The reason is that all function arguments are passed by reference
  ! * in Fortran. Setting P = zero(P1) set P1 to zero too, if these are
  ! * the same variable in the calling subroutine.
  ! * This can be avoided by creating temporary variables (inefficient)
  ! * or by copying the variable instead of setting it to zero;
  ! * but the only truly safe practice is the disallow allocated outputs.

This may be a temporary measure (can later replace zero with copy in these
routines, if this turns out a significant drawback in terms of efficiency).
At least, this way we don't have to deal with mysterious segmentation faults.
Derivative testing now runs, but illustrates that there is still something
wrong with the new derivative computations. This needs to be debugged asap.

Also, be very careful to NEVER FORGET to set the allocated attribute to
.FALSE. when it is defined in a data type. Otherwise, the compiler may set
it to .TRUE. on compilation, and that will certainly cause nasty bugs!