Following Gary's rules for the linear combinations now: we can overwrite input with output, but output has to be allocated before calling such a routine.
This works quite safely, but that however ruins the interface
operators such as +,-,*,/ - these are overloaded with functions,
and a function output is undefined. It can easily overwrite
the input, but it is not recognized as (un)allocated before using.
So avoiding the interface operators now, and replacing them all
with linComb and scMult subroutines.

The global code passes all tests; most mysteriously, it also
seems to pass the derivative test, even through the derivative
seems to have to opposite sign to what it should be. In my code,
I defined the data residuals as res=dat-psi, hence the derivative
should (I think!) be multiplied by +2 (not -2, as in Gary's code,
which uses res=psi-dat). I multiply by -2, and still the old
inversion works and the derivative tests are passed.

The new inversion produces correct numbers now (the RMS misfit 
is different from what I previously got because it is a sqrt
of my normalized sum of squares misfit). Still, it does not
converge! The misfit does not seem to decrease at all
(or significantly) as it runs (as if the gradient still has the
opposite sign somewhere...) DEBUGGING...

Need to check that nothing in broken in 2D and 3D MT, now that
I've made some changes to NLCG to ensure that the output of linComb
is always allocated. I haven't checked yet! So don't update the code
to this revision, if you want the MT codes to work.