IMPORTANT CHANGE
Merged Naser's MPI code to the trunk. I tried to do this is a way
that would allow us to propagate most of the changes to the stable
serial version.

I updated the Perl script fmkmf.pl with a quick & dirty fix:
if the module file name contains the string 'MPI', the use statement
is ignored, unless MPI is defined. This won't work for all cases,
but it will work as long as we name all MPI files consistently,
with an 'MPI' string included [unfortunately, couldn't concatenate
the lines such as #ifdef MPI; use MPI_main; #endif, since the use
statement can only follow a space in Fortran].

All of the configure scripts (and makefiles) are updated and work.
To allow for the #ifdef MPI statements, I had to use a preprocessor
with all compilers. This caused an issue with the include statements,
which was fixed by using #include statements instead (no spaces allowed).

This should work just like Naser's version in branches/MPI.
Having said that, Makefile3d.MPI still doesn't work for me on Lenz
(mpif.h not found). 
This is most likely a compiler flag / library problem. Would be nice
to figure out what is going on, to be able to test (and run) on Lenz.