Three minor updates.
1. Reverted the order of libraries for the PGI compiler,
to reflect the fact that the older version 6.1 accepts -llapack -lblas
(in that order!), but not -lblas -llapack. The newer versions
don't care about the order of these libraries.

2. Removed the intent(in) attributes from the count and write routines
for sensitivity matrix. Caused PGI 6.1 to bomb! Otherwise was correct.

3. Changed the default covariance parameters from 0.1 to 0.3 in 3D_MT.
This seems to work better in most 3D cases (one should use a covariance
file, anyway!)