The num. of polarization changes were even trickier than expected.. Fixed some bugs that were introduced in the previous commit (in some cases, a shortcut was previously taken and EMrhs_t was not created explicitly... this no longer works when an array of RHS_t values needs to be allocated).
Made some major changes to the internal structure of EMsolver, EMsolve3D
and modelOperator3D. There is now only one subroutine that initializes
model operators. It figures out what it needs to do from
1) a logical variable modelDataInitialized in modelOperator3D,
2) comparison between the saved and the new omega,
3) a logical variable "updated" in the model parameter (that gets set
to TRUE whenever the model parameter is changed in some way with a
routine from the model parameter module... we set it to FALSE when
the forward solver has been initialized). This is not the best solution!
The clarity of an honest comparison between 3D models is worth the extra
computations required to do that (I think). 
3D tested and works (inversion not tested).

EM solver should have a generic routine to create the boundary conditions
(in the form of EMrhs_t, NOT RHS_t).
Boundary conditions AND the model parameter should be passed as arguments
to EM solver routines, to avoid the current ambiguities. The forward modelling
in the EM solver SHOULD have the capability to work with a user-defined
forcing.
More work required before these changes are complete!!!