MAJOR CHANGES in the user interface.
1. The inversion now output the following files at every iteration:

*.rho - the output model m = Cm^{1/2} \hat{m} + m_0
*.prm - the output model parameter \hat{m} *before* smoothing 
*.imp - the output responses
*.res - the output weighted residuals

The user should be mostly interested in *.rho files (currently in WS format).
The files *.prm (also in WS format) could be used to restart the inversion
(see below).
The files *.imp and *.res are still in the legacy data format. The extensions
might change when the format is finally updated.

2. New command line arguments to run the inversion:

 Usage: -I NLCG rFile_Model rFile_Data [rFile_Cov lambda rFile_dModel]
 Here, lambda is the initial damping parameter
       rFile_dModel  is the starting rough model parameter perturbation
 OR
 Usage: -I NLCG rFile_Model rFile_Data [rFile_Cov rFile_invCtrl rFile_dModel]
 Here, rFile_invCtrl is the inversion control file in the format

 Model and data output file name    : Example
 Initial damping factor lambda      : 1.0
 Initial search step in model units : 1.0
 Restart when rms diff is less than : 2.0e-3
 Exit search when rms is less than  : 1.05
 Exit when lambda is less than      : 1.0e-4
 Maximum number of iterations       : 120

This implements a) the capability to restart with a *.prm model and b) the user
defined inverse solver configuration.

3. The output level of the program (i.e. how much output to screen and to files)
is set by the optional final argument -v [debug|full|regular|compact|result|none];
regular is of course the default.

4. The default file names in UserCtrl are now set to 'n' (rather than empty lines)
to avoid problems with inquire statements on IBM architecture. XLF compiler complains
if it encounters empty file names in inquire statements (checks for file existence).
Output file names should be longer than one character to be recognized... this could
be changed, but I think it's a reasonable expectation.

5. Changed the 3D MT forward solver output "finished solving etc" statements 
to standard output instead of standard error. Otherwise, these lines get lagged
or flushed out of order on clusters. Also, now output period instead of frequency...
makes it easier to look at.

The user documentation (that the user sees if he/she just types ./Mod3DMT) claims
that you can also specify a forward solver configuration file. This has not yet
been implemented in this commit. Will implement this option soon.