A major Update to backport the two-layer parallelization version from ModEMdevel to stable- Now it is possible to try out the finer-grain parallelization with   ModEM and PETSc (i.e. use more CPUs than the number of Transmitters). As I reserved the necessary interfaces, it is also possible now to build your own finer-grain parallel forward scheme upon the two-layer structure (and avoid the PETSc mess). See MPI/Main_MPI.f90 for details.
Well, to use the finer-grain parallel scheme in its current form, One needs to compile ModEM with proper configured PETSc lib. Please refer to http://mcs.anl. gov/petsc/ for the (exhausting) procedures for the setup of PETSc libs.

See CONFIG/Configure.SPETSc2.MAC.GFortran for a sample configuration file for the Makefile - Note that one needs to edit the '-pdir' parameter for the location of PETSc lib, in order to tell the makefile where to find the extra include and link files.

After the building, Simply assign more CPUs than nTx (nPer*nPol+1 in MT) to try the finer-grain parallel feature. One of the potential loopholes of the new version is still the dynamic load-balancing system. Although I have been using it for a while (seems robust to me so far), it might still produce some queer results in some situations I overlooked. As far as I tried, it is good enough for my “everyday work”, although the efficiency (for both time and memory) can certainly be improved.

Note that for the finer grain parallelization to work effectively, the problem to be solved should be large enough (in the sense of the degree of freedom). Testing the PETSc version with a toy model would most certainly lead to a slow-down as the benefit of more CPUs is easily overshadowed by the extra overhead introduced.

Another thing worth noting is that now ALL other versions (MF/SP/SP2) share the same two-layered parallel structure. But only the SPETSc2 version is able to effectively utilize it (non-PETSc programs will just behave like before).    So it IS also possible to assign more CPUs than "the necessary number" in those versions, albeit the extra CPUs won't do anything but wait while others do the job.

Let me know if you encounter any problems with this version!
Hao

Detailed change logs are as follows:

MPI/Main_MPI.f90
    updated to implement the new two-layered parallel structure 
    to utilize the (not really) new FWD_SPETSc2 forward routines.
    the program should behave similar to the previous structure if the provided
    number of worker CPUs is smaller or equal to the number of tasks 
    (i.e. Nworker <= nTx)
MPI/Main_MPI.bak
    backup of the previous Main_MPI.f90 - useful if one does not want to 
    tamper with the new two-layer parallel structure
3D_MT/ForwardSolver.f90
    added new optional mpi_comm parameters to FwdSolve and SensSolve for 
    the implement of two-layer parallel structure
2D_MT/ForwardSolver.f90
    added similar dummy optional mpi_comm parameters to FwdSolve and SensSolve
    to keep the compatibility of two-layer parallel structure
3D_MT/SolnSpace.f90
    fixed a bug when b%allocated== true but b%b(iTx)%allocated == false
    in create_rhsvector subroutine.
    well, normally this should not happen at all -
    this is more like a side effect when a dummy comb(b) is assigned 
    to one of the workers (non-leader) in a mpi subgroup. let's say after a
    REGROUP command, that worker becomes a leader. In this case the dummy comb
    has to be replaced by a real rhsvector. This could generate an error as 
    the dummy comb is not yet allocated (but treated like one as the programme
    only checked if b%allocated == true). 
3D_MT/FWD_SPETSc2/EMsolv3D.f90
    main driver module for the SPETSc2 forward branch, uses the external 
    PETSc libs to do the forward calculation in (finer grain) parallelization  
    the FWDsolve3D interface now has two actual subroutines: 3Ds and 3Dp
    FWDsolve3Ds is the old standalone forward code (without petsc)
    FWDsolve3Dp is the code with petsc parallelization
3D_MT/FWD_MF/EMsolv3D.f90
    added a dummy parameter in FWDsolve3D for the compatibility of the 
    two-layered parallel structure. 
3D_MT/FWD_SP/EMsolv3D.f90
    added a dummy parameter in FWDsolve3D for the compatibility of the 
    two-layered parallel structure. 
3D_MT/FWD_SP2/EMsolv3D.f90
    added a dummy parameter in FWDsolve3D for the compatibility of the 
    two-layered parallel structure. 
3D_MT/FWD_SPETSc2/modelOperator3D.f90
    essentially the same module as in the FWD_SP2 branch, might want to 
    keep only one copy of those 
3D_MT/FWD_SPETSc2/spModOpTest.f90
    essentially the same module as in the FWD_SP2 branch, might want to 
    keep only one copy of those 
3D_MT/FWD_SPETSc2/solver.f90
    obsolete if the finer grain parallelization is utilized
    essentially the same module as in the FWD_SP2 branch, might want to 
    keep only one copy of those 
3D_MT/FWD_SP2/modelOperator3D.f90
    removed the PETSc macro - no longer used as SPETSc2 has its own
    modelOperator module now.
CONFIG/Configure.SPETSc2.MAC.GFortran
    simple configure file for SPETSc2 - also have to modify the fmkmf.pl
    to use that (I just finally remembered why I didn't like perl).  
    NOTE: unfortunately the Makefile generated cannot be used gracefully by 
    typing "make", one has to use "make all" to do the compling and building
    (a silly PETSc include rule prevents you from doing that)
fmkmf.pl
    slightly changed to be compatible with the PETSc lib